+++ src/pages/EditDesign/Pattern/Pattern.css.ts
