+++ src/pages/EditDesign/Pattern/constants.ts
