+++ src/pages/EditDesign/components/DesignSizeImage/DesignSizeImage.css.ts
