+++ src/pages/EditDesign/components/FontSize/types.ts
