+++ src/pages/EditDesign/components/Footer/Footer.css.ts
