+++ src/pages/EditDesign/components/StepProgressBar/StepProgressBar.css.ts
