+++ src/routers/MyInformationRouter/index.tsx
@@ -3,13 +3,14 @@ import {
   MY_INFORMATION_CREATE_DESIGN_PATH,
   MY_INFORMATION_CREATE_PRODUCT_PATH,
 } from 'knitting/constants/path';
-import CreateDesign from 'knitting/pages/CreateDesign';
 import CreateProduct from 'knitting/pages/CreateProduct';
 import Error404 from 'knitting/pages/Error404';
 import MyInformation from 'knitting/pages/MyInformation';
 
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
+
+import CreateDesign from '../../pages/EditDesign/CreateDesign';
 
 const MyInformationRouter = (): React.ReactElement => {
   return (
